// src/enums/FormActionTypes.ts
export enum FormActionTypes {
    CheckAll = 'CHECK_ALL',
    ClearCheckBoxes = 'CLEAR_CHECKBOXES',
    UpdateAssignee = 'UPDATE_ASSIGNEE',
    UpdateCheckBoxes = 'UPDATE_CHECKBOXES',
    UpdateEndDate = 'UPDATE_END_DATE',
    UpdateFromDate = 'UPDATE_FROM_DATE',
    UpdateInterval = 'UPDATE_INTERVAL',
    UpdateIntervalCount = 'UPDATE_INTERVAL_COUNT'
}